+++ internal/framework/provider/provider.go
@@ -12,12 +12,13 @@ import (
 	providerschema "github.com/hashicorp/terraform-plugin-framework/provider/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	provideracl "github.com/hashicorp/terraform-provider-nomad/internal/framework/provider/acl"
+	"github.com/hashicorp/terraform-provider-nomad/internal/framework/provider/acl"
 )
 
 // Ensure NomadProvider satisfies various provider interfaces.
 var (
-	_ frameworkprovider.Provider = (*NomadProvider)(nil)
+	_ frameworkprovider.Provider                       = (*NomadProvider)(nil)
+	_ frameworkprovider.ProviderWithEphemeralResources = (*NomadProvider)(nil)
 )
 
 type NomadProvider struct {
@@ -137,7 +138,7 @@ func (p *NomadProvider) Configure(_ context.Context, _ frameworkprovider.Configu
 
 func (p *NomadProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
-		provideracl.NewACLAuthMethodResource,
+		acl.NewACLAuthMethodResource,
 	}
 }
 
@@ -146,5 +147,7 @@ func (p *NomadProvider) DataSources(_ context.Context) []func() datasource.DataS
 }
 
 func (p *NomadProvider) EphemeralResources(_ context.Context) []func() ephemeral.EphemeralResource {
-	return nil
+	return []func() ephemeral.EphemeralResource{
+		acl.NewIntroTokenEphemeralResource,
+	}
 }